requirements.txt
streamlit==1.28.0
PyMuPDF==1.23.7
openai==1.3.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.11.2
webdriver-manager==4.0.0
scikit-learn==1.3.2
torch==2.1.0
transformers==4.34.1
requests==2.31.0
